Restructured project files to take advantage of Spring Boot autoconfigurations and embed Tomcat and in-memory HSQLDB.
