--- src/main/webapp/index.jsp
@@ -1,5 +1,0 @@
-<html>
-<body>
-<h2>Hello World!</h2>
-</body>
-</html>
