+++ src/test/java/com/lamarjs/MainTests.java
@@ -1,0 +1,16 @@
+package com.lamarjs;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class MainTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
